+++ Framework/Utils/src/Utils.cxx
@@ -13,7 +13,7 @@
 ///
 /// \author Gabriele Gaetano Fronzé, gfronze@cern.ch
 
-#include "../include/Utils/Utils.h"
+#include "DPLUtils/Utils.h"
 
 namespace o2f = o2::framework;
