+++ Framework/Utils/test/test_RootTreeReader.cxx
@@ -16,7 +16,7 @@
 #include "Framework/InputSpec.h"
 #include "Framework/OutputSpec.h"
 #include "Framework/ControlService.h"
-#include "Utils/RootTreeReader.h"
+#include "DPLUtils/RootTreeReader.h"
 #include "Headers/DataHeader.h"
 #include "Headers/NameHeader.h"
 #include "../../Core/test/TestClasses.h"
