+++ Framework/Utils/test/test_RootTreeWriterWorkflow.cxx
@@ -17,8 +17,8 @@
 #include "Framework/OutputSpec.h"
 #include "Framework/ControlService.h"
 #include "Framework/CallbackService.h"
-#include "Utils/RootTreeWriter.h"
-#include "Utils/MakeRootTreeWriterSpec.h"
+#include "DPLUtils/RootTreeWriter.h"
+#include "DPLUtils/MakeRootTreeWriterSpec.h"
 #include "Headers/DataHeader.h"
 #include "../../Core/test/TestClasses.h"
 #include "FairMQLogger.h"
