+++ Steer/DigitizerWorkflow/src/TRDDigitWriterSpec.h
@@ -12,7 +12,7 @@
 #define STEER_DIGITIZERWORKFLOW_SRC_TRDDIGITWRITERSPEC_H_
 
 #include "Framework/DataProcessorSpec.h"
-#include "Utils/MakeRootTreeWriterSpec.h"
+#include "DPLUtils/MakeRootTreeWriterSpec.h"
 #include "Framework/InputSpec.h"
 #include "TRDBase/Digit.h"
