+++ Steer/DigitizerWorkflow/src/ZDCDigitWriterSpec.h
@@ -12,7 +12,7 @@
 #define STEER_DIGITIZERWORKFLOW_SRC_ZDCDIGITWRITERSPEC_H_
 
 #include "Framework/DataProcessorSpec.h"
-#include "Utils/MakeRootTreeWriterSpec.h"
+#include "DPLUtils/MakeRootTreeWriterSpec.h"
 #include "Framework/InputSpec.h"
 #include "ZDCSimulation/Digit.h"
 #include "ZDCSimulation/MCLabel.h"
