+++ cmake/O2Dependencies.cmake
@@ -373,14 +373,17 @@ o2_define_bucket(
 )
 
 o2_define_bucket(
-        NAME
-        DPLUtils_bucket
+    NAME
+    DPLUtils_bucket
 
-        DEPENDENCIES
-        O2FrameworkCore_bucket
-        Core
-        O2Headers
-        O2Framework
+    DEPENDENCIES
+    O2FrameworkCore_bucket
+    Core
+    O2Headers
+    O2Framework
+
+    INCLUDE_DIRECTORIES
+    ${CMAKE_SOURCE_DIR}/Framework/Utils/include
 )
 
 o2_define_bucket(
